refactor: rename finalise to finalize
