+++ lib/workers/repository/finalize/prune.ts
